BUGFIX: Resolves scikit x numpy x pandas issue
Pandas v2.+ has now allows a new null type, 'pd.NaT', which is not the
same than ol' 'np.NaN', and is not recognized as NaN by Scikit.

To solve this, the dataframe is passed to
`df = df.astype(dtype='float64', errors='ignore')`

This may be an issue for future categorical data, but may be ignored as
we currently don't include it.

Also, the `Int8` type is not the same that `np.int`, so we also
default to it. Should suggest a fix for the Scikit part instead of the
general pre-processing one.

PD1: the df index now includes the 'pacientes' col.
PD2: changes the 'warning.ignore' as to avoid a Numba deprecation thingy